requirements.txt
# ----------------------
# Flask
# ----------------------
Flask
Flask-Login
Flask-Migrate
Flask-OpenID
Flask-SQLAlchemy
Flask-Script
Flask-WTF
Jinja2
Mako
MarkupSafe
SQLAlchemy
WTForms
Werkzeug
alembic
argparse
funcsigs
gunicorn
itsdangerous

pbr
psycopg2
python-editor
python-openid
BeautifulSoup4
requests
six
wsgiref

# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------

